server/requirements.txt
numpy==1.19.4
pandas==1.2.0
scikit-learn==0.24.0
cloudpickle==1.6.0
python-dotenv==0.15.0
fastapi==0.63.0
requests==2.25.1
uvicorn==0.13.0
prometheus-fastapi-instrumentator==5.7.1
pytest==6.2.2
isort==5.7.0
black==20.8b1
autoflake==1.4